Moves duplicated code into a test helper method
